Move auth button from HeaderNavMenu to Header
